Add files via upload
This code contains two methods of extracting data from the UK Company House dataset, which is updated each month, and contains around 5m records. In each case the data is filtered. 
This involves downloading and filtering large zip files with fields such as company name, address, status, postcode, SIC code.
In Method 1, filtering is by town. However, due to inconsistencies in the data, this method was less reliable.
In Method 2, filtering is by finding companies with the first parts of the postcode e.g. BH2 or B54.
The data is sent to a CSV export in Google Colabs.